Userviews added
Added a userviews table that tracks which thread the user has recently
viewed and at which time. This is used by the frontend to indicate which
threads should be bolded to show the user that this thread either has a
new post since the last time they visited or they have not seen that
thread before.