Moved userviews to users.thread_views
The userviews table was deleted and replaced with a new table called
thread_views. This new table is namespaced under a new schema called,
'users'. This required that a new schema sql file was created for just
the 'users' namespace. All the creation files were updated to reflect
this change.

After the edits to the sql files, the users.js file was also updated to
reflect the change in name.