Attached users to posts on byThread
The byThread call returns all the posts for a thread but did not tie the
user to the post at all. Thus posts had a user_id but no username or
any other info. This commit at least binds the user_id and the username
to the post under the user property.