Create/Import method merge in threads/posts
Merged the post insert opertations from the import and create
functions in both the index.js files for threads and post. The code for
import and create posts/threads have reached a point where they are
doing nearly the exact same thing. This commit pull the duplicate parts
of the two functions into one function. Then the import and create
functions merely pass in the differences into the merged function. This
was done for both posts and threads.