board_moderators table, return moderators for boards
A new table called board_moderators was created with a migration that
is a many to many table of users to boards. This list is the list of
moderators for any particular board.

The boards.find and boards.allCategories methods were updated to return
the moderators information for each board.