Bug fixes, refacor users code again
Fixed a few bugs related to the last refactor. The biggest being, not
being able to insert a new fields property for users.profiles due to a
lack of a fields property on the original copy. Also, the userByUsername
has been reverted to return undefined if a username is not found. The
import method has also been updated to support the fields property. The
last bug fix revolved around always returning Promise.rejects because
they do not stop the flow of control otherwise.