Post count metadata, optimized threads.find and byBoard
Added a column to the metadata.threads table that keeps track of the
number of posts in a thread. This is a quicker lookup compared to
counting a thread with a large number of posts.

The threads.find and thread.byBoard methods were also updated to be
quicker by moving more the work into the SQL queries rather than doing
code pivots that keep going back to the DB.