Refactor libs, posts
Removed unused libs from most models and moved all test based libs to
the dev section of the package.json file.

The posts/index.js file was refactored to move all timestamps into the
postgres. The only timestamps not moved over are the ones that come in
through the import method.