Fixed issues with tests
The category properties were not being set properly due to using the
wrong variable name.

The thread tests failed due to a leak in the compressed UUIDs which
caused the expectations function to fail. It was trying to compare a
uncompressed UUID to a compressed UUID. This may be something to look
into at a later time but does not matter so much since the leak is only
possible with a value that is passed into the create method. In any
case, it is fixed for thread.create but persists in other methods.

The slugTransform method was edited to handle an array of UUIDs.