Bug Fix: deslugify threadId on thread view
There was one area where the threadId was not being deslugified
correctly. The method to update or insert a new thread view is pulling
the threadId on the fly from the user viewing the thread and when the
threadId is pulled, it needs to be deslugified or else postgres will
through an error.