Bug Fix: Saving user thread views
There was an issue saving a user viewing a thread. The view itself was
being saved but the logs for that user would return with the ids
deslugified. This made it difficult for the frontned to match up the
correct thread and thus the threads would always appear as not viewed.