Thread sticky
Implemented thread sticky on the DB side of the code. This adds a column
to the threads table that designates whether a thread is a sticky thread
or not and if should be pulled out of the regular paging rotation.

A method was also created dedicated specifically to set the sticky value
for a specifci thread.