Updated thread test for sticky threads
With the introduction of sticky threads, the output of the byBoard
method has changed. Instead of being just an array of threads, it now
returns an object with two arrays, one filled with normal threads and
one filled with the sticky threads.

The threads fake and test file were updated to reflect this change.