Added a helper update function
Added a function to the helper.js file that simplifies the process of
updating a object property. This function checks if the incoming value
is an empty string, and sets the value to null. If the incoming value is
undefined, the old value is used instead.

This function cleans up the update code in most models.