Transactions for threads model
Updated all the modifying method in the index.js file for threads to
utilize Postgres transactions where appropriate. Thread lock and sticky
are not using transactions because they are one line queries that are
already inherently wrapped in a transaction block. The rest of the calls
explicitly create a transaction to run their queries.