Transaction for users model
Updated some of the modifying methods in the index.js file for the users
to utilize Postgres transactions. The methods updated are the import,
create, update, insert/updateUserProfile, and update thread view methods.
There are other methods that need to be address but will but pushed off
into another branch. The remaining methods are read only methods and do
not need transactions.