Board mapping intersection table
The way boards are retrieved for the front page of the forum has been
overhauled to use a hierarchical intersection table to keep track of
where each board belongs. Both the updateCategories method and the
allCategories method have been updated to use this new table. This new
way of handling board hierarchy also allows for better support for child
boards.

A migration was added to create this new table as well as delete all the
parent_board_id, children_ids, and category_id columns from the boards
tables.