Throw error when moving thread to current board
The threads.move method now handles the edge case where the user is
asking to move the board to the same board the thread is already in.
This actually causes a deadlock in the DB since the threads.move method
tries to lock but the current board and the new board. After locking the
thread, the method now checks to see if the current board id matches the
new board id, if so, an Promise.OperationalError is thrown.