Soft Delete migration
A migration file that adds the deleted column to all the major DB
models: posts, threads, boards, and users. This 'deleted' column is of
type boolean and mainly acts as a flag to signify that the particular
instance has been soft deleted. This soft delete will keep the instance
data in the DB but cannot be viewed through the normal DB methods like
find, byThread, byBoard, etc.