Bug Fix: deleted threads, purge threads
Fixed two bugs dealing with deletion. The first bug covers deleted
threads. If a thread is marked deleted, it is still visible during the
tail end of the posts.deleted, undelete, and purge methods. It then
assigns the deleted thread as the most recently posted thread in the
board's metadata. This has been fixed by filtering out deleted threads
in the query to update the board's metadata.

The other bugs deals with a post that was deleted, then purged. There
was nothing stopping the purge method from decrementing the thread and
board metadata for a second time after the post was deleted. There is
now a check put into place for each of the metadata update queries
checking to see if the post was already deleted.