Thread Deletion
Added the delete, undelete, and purge methods to handle thread deletion
in both soft and hard modes. The delete and undelete methods flags the
thread appropriately while also updating all relevant metadata. The
purge method permanently removes the thread from the DB while also
removing all other rows linked to it (posts, metdata.threads).

The retrieve-type methods were all updated to handle the deleted flag
while either retrieving a specific thread or a collection of threads by
board.

The threads.move method was also updated to account for the situation
where a thread no longer exists when this method is called. It will now
return a DeletionError if a thread cannot be found.