Board Deletion, minor fixes
Added the delete method to handle a hard deletion of a board object from
the database. The board, thread, and post models have all been updated
with a method that allows each model to check whether the board their
belong to is visible from the board_mapping table.

Minor Fixes: All the find methods now return an error if the item being
looking for is not found.