Thread Deletion update
Threads are no longer soft deleted. They can only be moved to a board
that is not publicly viewable. The delete method has been replaced with
a method that does a hard delete instead. This method is not advised
right now because it does not properly handle user post count updates.