Post Deletion update and User Deletion
Post deletion was updated to no longer update the post counts among the
various metadata. Instead, it simply marks the post as being deleted. In
turn, all the read type methods now pull in the deleted posts but also
pulls in the post.deleted property. This signifies that the post has
been deleted and the cleaning of the post is handled at a higher level.

User Deletion has also been added must the same way post deletion is
handled. All read type methods check to see if a post's user has been
deactivated and passes along as flag to mark that the user is not
active. This flag is set to post.user.deleted.