Image expiration and post relations
The image expiration table has been moved from memdown on the server
side to postgres at the DB layer. This allows for the persistents of
images expirations.

A new table called images_posts was also created to log images are being
reference in all posts. This allows the frontend to query for images not
being used anymore and removes them from the cdn of choice.