Private Messages
A new migration was added that creates two tables in the database where
the conversation model and the private message model are being stored.

The necessary db methods were created to manipulate those tables. These
method include creating and deleting the models as well as retriving a
list of the most recent messages and the most recent message for any
given conversation. A few other helper methods were also created such as
a method that determines if a user is in a conversation or if a user is
the person that sent the message.