Support for post positions
Added a migration that adds a position column to posts that keeps track
of the position of the post relative to it's thread and created_at
timestamp.

This column should be updated on it's own post row creation and if any
posts with a position lower than itself is modified.

The posts.byThread method was also modified to run strictly on the post
position to search for posts. All other methods that return a post now
include their position as well for easier recall.