Added support for board last post position
Created a migration that adds a column to the metadata.boards table for
the post position so that in the frontend, a quick link can be built to
send users to the last page of a thread.

The migration also updated the delete_thread and update_thread DB
functions to update the last post data accordingly with the new column.