Bug Fix: post position triggers, last post data
Two bugs were fixed. The username for the board's last post data was set
to the username of the user that started the thread and not the last
person that made a post in that thread. This has been fixed to use the
correct username.

The post position was also not being set correctly. It was being set to
the last post position - 1. It has now been fixed by moving the order of
operations in the create_post and delete_post trigger such that the
position of the post is set before the last post data is set.

The posts.byThread was also adjusted because it was returning the last
post of the previous page with any newer pages. The algorithm to
determine the page from the start parameter now accounts for that
mistake.