Support for roles as a model
Roles are now a model and thus have their own folder at the top level.
This commit adds support for adding and removing roles. It also supports
updating the permissions of a role. Lastly, there is support for getting
all roles at once. The roles model was also added to the top index file
so that the roles method are accessible from the db object returned to
the caller of this library.