Board Moderators implementations
A migration was created to enfore the cascade relationship between the
boards and user in board_moderators. Now whenever a board or user is
deleted, the corresponding rows in the board_moderators table will be
removed as well. A unique index on board_id to user_id was also created
to ensure duplicates don't exists in the table too.

The boards/index.js file was updated to pull the board moderators for
each board requested. This is done in the initial query rather than
querying for the board moderators after a board is pulled out for
performance sake.

The posts/index.js file was updated to return the post's board as well
as a boolean to designate if the board for that post is visible. This
fixes the issue concerning hiding post from uncategorized boards in the
user profile. This also allows the pre method on the frontend to
intelligently hide posts if a user is deleted by the viewer is a mod.
Only the posts that are within the boards the viewer is a moderator of,
will be shown to the user.

A new folder with db methods specifically for moderators has been added
that allows for adding and remove mods from a board as well as checking
if a user is a moderator of a board via board_id, thread_id, and
post_id. One other method was added returning all the boards a user
moderates. The index file was updated to include this folder.

Lastly, the helper file was update to return 'undefined' for any input
into the slugify/deslugify methods that resembles an empty string. This
cuts down on intermediate errors along the way to the db.