Refactoring threads.byBoard
The entire threads.byBoard method was refactored to be much faster. It
now splits the retrieval of the normal and sticky threads using a
Promise.join. This allows both queries to run at the same time rather
than one after the other. Also, the threadLastPost and an extra query to
get the last thread view, next unread post id, and next unread post
position was added to the normal and stick thread retrieval. Those two
queries were originally run after the retrieval and the extra calls back
to the DB incurred slight overhead. Now with the everything rolled into
one query, the call is much faster.

With the two queries being rolled into the original retrieval, all
duplicate code was removed from threads and users. The helper function
was also updated with new keywords to filter by.