DB migrations for thread_views, fixes
Fixed an issue with calling the putThreadViews method for users. The
time column was not being declared although a time value was being
passed into the query.

Created a new migration to enfore a unique relationship between the
user_id and thread_id in the users.thread_views table and dropped the id
column because it doesn't really do much.