New Watchlist Folder
Created a new folder called watchlist that holds all the db methods
necessary to implement the watchlist feature. This includes grabbing all
threads and/or boards a user is watching, getting a list of the most
recent unread threads, and getting a list of the most recent threads
regardless of whether they are unread or not.

This allows the user to view unread threads they are interested in,
while also viewing all their thread/board subscriptions. They are also
allowed to edit their watchlist by removing or adding threads and boards
individually.