postable_by field on boards, categories
This allows the system to track which boards and categories let a groups
of users post to that model. This works similar to the viewable_by flag
that allows a group of users to view a board or not. It also works the
same way in the sense that all users with a higher priority are also
allowed to view/post. (Higher priority = lower numerical value).