Bug Fix: User notes break when viewer has more than one role
This fixes a bug where if the admin/moderator viewer of a profile has
more than one role, the subquery to grab all the user note info breaks
because more than one row is return for that particular subquery.

This fixes the problem by adding a ORDER BY clause to the sub query and
a 'LIMIT 1' clause to ensure that only the highest priority is returned
for this subquery.