Migration: fix for "portal_board_id" in previous migrations
The previous migration had an error in adding a column to the
configrations table. It added "portal_board_id" where it should have
been "portal.board_id". This migration removes "portal_board_id" if it
exists and replaces it with the proper "portal.board_id" column. This
may require a manual update to the column's value itself in the admin
panel if it was previously set.